requirements.txt
asgiref==3.2.3
Django==2.1.15
django-annoying==0.10.6
django-filter==2.2.0
django-jsonstore==0.4.1
django-material==1.6.3
django-viewflow==1.6.0
factory-boy==2.12.0
Faker==4.0.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
